Move files from unversioned project
The development of this application was started on another project
structure and without version control.

For this moment - the application builds a bracket with competitors
placed on it and coordinates assigned to each of competitor cell.
A list of competitors is loaded from a text file. The size of the bracket
is calculated automatically based on a number of competitors.