Modify SpreadsheetExporter to write each bracket into a separate file
Because sheet name length is limited to 31 characters for xls files and
category names are usually longer and cannot be shortened, it is necessary
to write each bracket into a separate file.
When the sheet name is longer, it is being cut to 31 characters.
Many category names are almost identical, differ only with a several
characters at the end - so after cutting we get multiple categories with
the same name. It caused a program crush during the sheet creation, with
error message saying that a sheet with the given name already exists.
It is necessary to have a full category name - it cannot be trimmed.