Improve zooming mechanism
Bracket grid pane scale is now retained while browsing through brackets.
Brackets positions are also correct (aligned top - left corner).
Scroll pane resizes automatically when zoom of the bracket changes.

To achieve correct behavior of zoomed brackets it was necessary to wrap
the bracketGridPane with Group layout. However, wrapping grid pane in a
group layout caused performance issues while displaying larger brackets
(it was noticeable starting from 32-competitor brackets) each time the
bracket was displayed.
As a workaround for this problem it was decided to draw each bracket on
its own grid pane and while browsing - just switch between grid panes.
It did not fully resolved the performance problem. When a bracket is being
displayed for the first time, a delay can bee seen. But when it is a second
time and further - no delays are observed.